Restructure image_measures hierarchy
- Move the three measures to a single module
- Make it a top-level module
- Delete the sub package
