+++ image_measures.py
@@ -135,3 +135,42 @@ def _measure(num_objs, sum_notnull):
     sum_vals = np.sum(num_objs)
     measure_value = 1 - float(sum_vals) / (sum_notnull * nlevels)
     return measure_value
+
+
+def isotope_pattern_match(images_flat, theor_iso_intensities):
+    """
+    This function calculates a match between a list of isotope ion images and a theoretical intensity vector.
+
+    :param images_flat: 2d array of pixel intensities with shape (d1, d2) where d1 is the number of images and d2 is the number of pixels per image, i.e. :code:`images_flat[i]` is the i-th flattened image
+    :param theor_iso_intensities: 1d array of theoretical isotope intensities with shape d1, i.e :code:`theor_iso_intensities[i]` is the theoretical isotope intensity corresponding to the i-th image
+    :return: measure value between 0 and 1, bigger is better
+    :rtype: float
+    """
+    image_ints = []
+    not_null = images_flat[0] > 0
+    for ii, _ in enumerate(theor_iso_intensities):
+        image_ints.append(np.sum(images_flat[ii, not_null]))
+    pattern_match = 1 - np.mean(abs(theor_iso_intensities / np.linalg.norm(theor_iso_intensities) -
+                                    image_ints / np.linalg.norm(image_ints)))
+    if pattern_match == 1.:
+        return 0
+    return pattern_match
+
+
+def isotope_image_correlation(images_flat, weights=None):
+    """
+    Function for calculating a weighted average measure of image correlation with the principle image.
+
+    :param images_flat: numpy array of pixel intensities
+    :param weights: weighting to put on each im
+    :return: measure_value
+    """
+
+    if len(images_flat) < 2:
+        return 0
+    else:
+        # slightly faster to compute all correlations and pull the elements needed
+        iso_correlation = np.corrcoef(images_flat)[1:, 0]
+        # when all values are the same (e.g. zeros) then correlation is undefined
+        iso_correlation[np.isinf(iso_correlation)] = 0
+        return np.average(iso_correlation, weights=weights)
